// frontend/src/api/generated/models/review.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Conecta Bairro API
 * API for connecting professionals with clients in a curated marketplace
 * OpenAPI spec version: 1.0.0
 */

export interface Review {
  id?: string;
  userName?: string;
  userAvatarUrl?: string;
  rating?: number;
  comment?: string;
  createdAt?: string;
}